Cloned CleverHans directly since pip version is outdated. Created a template to use cleverhans attacks in cleverHans.py
