+++ src/main.py
@@ -47,4 +47,4 @@ def run(args) -> None:
 
 
 if __name__ == "__main__":
-    run(None)
+    main()
